requirements.txt
pyyaml
bcrypt
python-telegram-bot
Flask
requests
pytest
flake8